Updated Data Binding

+++ app/src/main/java/com/example/a160419095_advancenativeuts/view/BookDetailFragment.kt
@@ -25,7 +25,7 @@ class BookDetailFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         viewModel = ViewModelProvider(this).get(BookDetailModel::class.java)
         val id = BookDetailFragmentArgs.fromBundle(requireArguments()).bookId
-        viewModel.detail(id)
+        viewModel.detail(id.toInt())
 
         observeViewModel()
     }
